apps/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
requests==2.32.3

# LLM Integration
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.0.0